+++ src/spyglass/position/v1/position_dlc_pose_estimation.py
@@ -86,9 +86,10 @@ class DLCPoseEstimationSelection(SpyglassMixin, dj.Manual):
 
         Parameters
         ----------
-        key: DataJoint key specifying a pairing of VideoRecording and Model.
-        task_mode (bool): Default 'trigger' computation.
-        Or 'load' existing results.
+        key: dict
+            DataJoint key specifying a pairing of VideoRecording and Model.
+        task_mode: bool, optional
+            Default 'trigger' computation. Or 'load' existing results.
         params (dict): Optional. Parameters passed to DLC's analyze_videos:
             videotype, gputouse, save_as_csv, batchsize, cropping,
             TFGPUinference, dynamic, robust_nframes, allow_growth, use_shelve
@@ -231,8 +232,10 @@ class DLCPoseEstimation(SpyglassMixin, dj.Computed):
                 dlc_result.creation_time
             ).strftime("%Y-%m-%d %H:%M:%S")
 
-            logger.logger.info("getting raw position")
-            interval_list_name = (
+            # get video information
+            _, _, meters_per_pixel, video_time = get_video_path(key)
+            # check if a position interval exists for this epoch
+            if interval_list_name := (
                 convert_epoch_interval_name_to_position_interval_name(
                     {
                         "nwb_file_name": key["nwb_file_name"],
@@ -240,16 +243,16 @@ class DLCPoseEstimation(SpyglassMixin, dj.Computed):
                     },
                     populate_missing=False,
                 )
-            )
-            spatial_series = (
-                RawPosition()
-                & {**key, "interval_list_name": interval_list_name}
-            ).fetch_nwb()[0]["raw_position"]
-            _, _, _, video_time = get_video_path(key)
-            pos_time = spatial_series.timestamps
-            # TODO: should get timestamps from VideoFile, but need the video_frame_ind from RawPosition,
-            # which also has timestamps
-            key["meters_per_pixel"] = spatial_series.conversion
+            ):
+                logger.logger.info("Getting raw position")
+                spatial_series = (
+                    RawPosition()
+                    & {**key, "interval_list_name": interval_list_name}
+                ).fetch_nwb()[0]["raw_position"]
+            else:
+                spatial_series = None
+
+            key["meters_per_pixel"] = meters_per_pixel
 
             # Insert entry into DLCPoseEstimation
             logger.logger.info(
@@ -275,15 +278,17 @@ class DLCPoseEstimation(SpyglassMixin, dj.Computed):
             idx = pd.IndexSlice
             for body_part, part_df in body_parts_df.items():
                 logger.logger.info("converting to cm")
+                key["analysis_file_name"] = AnalysisNwbfile().create(  # logged
+                    key["nwb_file_name"]
+                )
                 part_df = convert_to_cm(part_df, meters_per_pixel)
                 logger.logger.info("adding timestamps to DataFrame")
                 part_df = add_timestamps(
-                    part_df, pos_time=pos_time, video_time=video_time
+                    part_df,
+                    pos_time=getattr(spatial_series, "timestamps", video_time),
+                    video_time=video_time,
                 )
                 key["bodypart"] = body_part
-                key["analysis_file_name"] = AnalysisNwbfile().create(
-                    key["nwb_file_name"]
-                )
                 position = pynwb.behavior.Position()
                 likelihood = pynwb.behavior.BehavioralTimeSeries()
                 position.create_spatial_series(
@@ -291,8 +296,10 @@ class DLCPoseEstimation(SpyglassMixin, dj.Computed):
                     timestamps=part_df.time.to_numpy(),
                     conversion=METERS_PER_CM,
                     data=part_df.loc[:, idx[("x", "y")]].to_numpy(),
-                    reference_frame=spatial_series.reference_frame,
-                    comments=spatial_series.comments,
+                    reference_frame=getattr(
+                        spatial_series, "reference_frame", ""
+                    ),
+                    comments=getattr(spatial_series, "comments", "no commwnts"),
                     description="x_position, y_position",
                 )
                 likelihood.create_timeseries(
@@ -329,6 +336,7 @@ class DLCPoseEstimation(SpyglassMixin, dj.Computed):
                     analysis_file_name=key["analysis_file_name"],
                 )
                 self.BodyPart.insert1(key)
+                AnalysisNwbfile().log(key, table=self.full_table_name)
 
     def fetch_dataframe(self, *attrs, **kwargs):
         entries = (self.BodyPart & self).fetch("KEY")