+++ src/spyglass/position/v1/position_dlc_position.py
@@ -1,3 +1,5 @@
+from time import time
+
 import datajoint as dj
 import numpy as np
 import pandas as pd
@@ -167,6 +169,7 @@ class DLCSmoothInterp(SpyglassMixin, dj.Computed):
             path=f"{output_dir.as_posix()}/log.log",
             print_console=False,
         ) as logger:
+            AnalysisNwbfile()._creation_times["pre_create_time"] = time()
             logger.logger.info("-----------------------")
             idx = pd.IndexSlice
             # Get labels to smooth from Parameters table
@@ -224,7 +227,7 @@ class DLCSmoothInterp(SpyglassMixin, dj.Computed):
                 .fetch_nwb()[0]["dlc_pose_estimation_position"]
                 .get_spatial_series()
             )
-            key["analysis_file_name"] = AnalysisNwbfile().create(
+            key["analysis_file_name"] = AnalysisNwbfile().create(  # logged
                 key["nwb_file_name"]
             )
             # Add dataframe to AnalysisNwbfile
@@ -267,6 +270,7 @@ class DLCSmoothInterp(SpyglassMixin, dj.Computed):
             )
             self.insert1(key)
             logger.logger.info("inserted entry into DLCSmoothInterp")
+            AnalysisNwbfile().log(key, table=self.full_table_name)
 
     def fetch1_dataframe(self):
         nwb_data = self.fetch_nwb()[0]