+++ src/spyglass/utils/dj_merge_tables.py
@@ -1,8 +1,7 @@
-import re
-from contextlib import nullcontext
 from inspect import getmodule
 from itertools import chain as iter_chain
 from pprint import pprint
+from re import sub as re_sub
 from time import time
 from typing import Union
 
@@ -10,7 +9,7 @@ import datajoint as dj
 from datajoint.condition import make_condition
 from datajoint.errors import DataJointError
 from datajoint.preview import repr_html
-from datajoint.utils import from_camel_case, to_camel_case
+from datajoint.utils import from_camel_case, get_master, to_camel_case
 from IPython.core.display import HTML
 
 from spyglass.utils.logging import logger
@@ -18,6 +17,36 @@ from spyglass.utils.logging import logger
 RESERVED_PRIMARY_KEY = "merge_id"
 RESERVED_SECONDARY_KEY = "source"
 RESERVED_SK_LENGTH = 32
+MERGE_DEFINITION = (
+    f"\n    {RESERVED_PRIMARY_KEY}: uuid\n    ---\n"
+    + f"    {RESERVED_SECONDARY_KEY}: varchar({RESERVED_SK_LENGTH})\n    "
+)
+
+
+def is_merge_table(table):
+    """Return True if table fields exactly match Merge table."""
+
+    def trim_def(definition):
+        return re_sub(
+            r"\n\s*\n", "\n", re_sub(r"#.*\n", "\n", definition.strip())
+        )
+
+    if isinstance(table, str):
+        table = dj.FreeTable(dj.conn(), table)
+    if not isinstance(table, dj.Table):
+        return False
+    if get_master(table.full_table_name):
+        return False  # Part tables are not merge tables
+    if not table.is_declared:
+        if tbl_def := getattr(table, "definition", None):
+            return trim_def(MERGE_DEFINITION) == trim_def(tbl_def)
+        logger.warning(
+            f"Cannot determine merge table status for {table.table_name}"
+        )
+        return True
+    return table.primary_key == [
+        RESERVED_PRIMARY_KEY
+    ] and table.heading.secondary_attributes == [RESERVED_SECONDARY_KEY]
 
 
 class Merge(dj.Manual):
@@ -34,32 +63,21 @@ class Merge(dj.Manual):
         super().__init__()
         self._reserved_pk = RESERVED_PRIMARY_KEY
         self._reserved_sk = RESERVED_SECONDARY_KEY
-        merge_def = (
-            f"\n    {self._reserved_pk}: uuid\n    ---\n"
-            + f"    {self._reserved_sk}: varchar({RESERVED_SK_LENGTH})\n    "
-        )
         if not self.is_declared:
-            # remove comments after # from each line of definition
-            if self._remove_comments(self.definition) != merge_def:
+            if not is_merge_table(self):  # Check definition
                 logger.warn(
-                    "Merge table with non-default definition\n\t"
-                    + f"Expected: {merge_def.strip()}\n\t"
-                    + f"Actual  : {self.definition.strip()}"
+                    "Merge table with non-default definition\n"
+                    + f"Expected:\n{MERGE_DEFINITION.strip()}\n"
+                    + f"Actual  :\n{self.definition.strip()}"
                 )
             for part in self.parts(as_objects=True):
                 if part.primary_key != self.primary_key:
-                    logger.warn(
+                    logger.warn(  # PK is only 'merge_id' in parts, no others
                         f"Unexpected primary key in {part.table_name}"
                         + f"\n\tExpected: {self.primary_key}"
                         + f"\n\tActual  : {part.primary_key}"
                     )
         self._source_class_dict = {}
-
-    def _remove_comments(self, definition):
-        """Use regular expressions to remove comments and blank lines"""
-        return re.sub(  # First remove comments, then blank lines
-            r"\n\s*\n", "\n", re.sub(r"#.*\n", "\n", definition)
-        )
 
     @staticmethod
     def _part_name(part=None):
@@ -121,9 +139,6 @@ class Merge(dj.Manual):
         """
 
         cls._ensure_dependencies_loaded()
-
-        if not restriction:
-            restriction = True
 
         # Normalize restriction to sql string
         restr_str = make_condition(cls(), restriction, set())
@@ -332,6 +347,9 @@ class Merge(dj.Manual):
                             + f"{part_name}:\n\tData:{row}\n\t{keys}"
                         )
                     key = keys[0]
+                    if part & key:
+                        print(f"Key already in part {part_name}: {key}")
+                        continue
                     master_sk = {cls()._reserved_sk: part_name}
                     uuid = dj.hash.key_hash(key | master_sk)
                     master_pk = {cls()._reserved_pk: uuid}
@@ -351,22 +369,12 @@ class Merge(dj.Manual):
                 part.insert(part_entries, **kwargs)
 
     @classmethod
-    def _safe_context(cls):
-        """Return transaction if not already in one."""
-        return (
-            cls.connection.transaction
-            if not cls.connection.in_transaction
-            else nullcontext()
-        )
-
-    @classmethod
     def _ensure_dependencies_loaded(cls) -> None:
         """Ensure connection dependencies loaded.
 
         Otherwise parts returns none
         """
-        if not dj.conn.connection.dependencies._loaded:
-            dj.conn.connection.dependencies.load()
+        dj.conn.connection.dependencies.load()
 
     def insert(self, rows: list, **kwargs):
         """Merges table specific insert, ensuring data exists in part parents.
@@ -761,7 +769,7 @@ class Merge(dj.Manual):
                 "No merge_fetch results.\n\t"
                 + "If not restricting, try: `M.merge_fetch(True,'attr')\n\t"
                 + "If restricting by source, use dict: "
-                + "`M.merge_fetch({'source':'X'})"
+                + "`M.merge_fetch({'source':'X'}"
             )
         return results[0] if len(results) == 1 else results
 
@@ -777,20 +785,26 @@ class Merge(dj.Manual):
 
     def delete(self, force_permission=False, *args, **kwargs):
         """Alias for cautious_delete, overwrites datajoint.table.Table.delete"""
-        for part in self.merge_get_part(
-            restriction=self.restriction,
-            multi_source=True,
-            return_empties=False,
+        if not (
+            parts := self.merge_get_part(
+                restriction=self.restriction,
+                multi_source=True,
+                return_empties=False,
+            )
         ):
+            return
+
+        for part in parts:
             part.delete(force_permission=force_permission, *args, **kwargs)
 
-    def super_delete(self, *args, **kwargs):
+    def super_delete(self, warn=True, *args, **kwargs):
         """Alias for datajoint.table.Table.delete.
 
-        Added to support MRO of SpyglassMixin"""
-        logger.warning("!! Using super_delete. Bypassing cautious_delete !!")
-
-        self._log_use(start=time(), super_delete=True)
+        Added to support MRO of SpyglassMixin
+        """
+        if warn:
+            logger.warning("!! Bypassing cautious_delete !!")
+            self._log_delete(start=time(), super_delete=True)
         super().delete(*args, **kwargs)
 
 
@@ -802,40 +816,21 @@ _Merge = Merge
 
 def delete_downstream_merge(
     table: dj.Table,
-    restriction: str = None,
-    dry_run=True,
-    recurse_level=2,
-    disable_warning=False,
     **kwargs,
 ) -> list:
     """Given a table/restriction, id or delete relevant downstream merge entries
 
-    Parameters
-    ----------
-    table: dj.Table
-        DataJoint table or restriction thereof
-    restriction: str
-        Optional restriction to apply before deletion from merge/part
-        tables. If not provided, delete all downstream entries.
-    dry_run: bool
-        Default True. If true, return list of tuples, merge/part tables
-        downstream of table input. Otherwise, delete merge/part table entries.
-    disable_warning: bool
-        Default False. If True, don't warn about restrictions on table object.
-    kwargs: dict
-        Additional keyword arguments for DataJoint delete.
-
-    Returns
-    -------
-    List[Tuple[dj.Table, dj.Table]]
-        Entries in merge/part tables downstream of table input.
+    Passthrough to SpyglassMixin.delete_downstream_merge
     """
+    logger.warning(
+        "DEPRECATED: This function will be removed in `0.6`. "
+        + "Use AnyTable().delete_downstream_merge() instead."
+    )
+
     from spyglass.utils.dj_mixin import SpyglassMixin
 
     if not isinstance(table, SpyglassMixin):
         raise ValueError("Input must be a Spyglass Table.")
     table = table if isinstance(table, dj.Table) else table()
 
-    return table.delete_downstream_merge(
-        restriction=restriction, dry_run=dry_run, **kwargs
-    )
+    return table.delete_downstream_merge(**kwargs)